added date selection in food diary. currently only add to current date (based on utc time).

+++ app/routes.py
@@ -1,12 +1,12 @@
-from flask import Flask, render_template, request, url_for, flash, redirect
+from flask import Flask, render_template, request, url_for, flash, redirect, session
 from app import app, db
 import requests
 from jinja2 import Template
-from app.forms import SearchForm, LoginForm, RegistrationForm, AddToDiaryForm, RemoveFood
+from app.forms import SearchForm, LoginForm, RegistrationForm, AddToDiaryForm, RemoveFood, DiaryDatePicker
 from flask_login import login_required, logout_user, current_user, login_user
 from werkzeug.urls import url_parse
 from app.models import User, Food
-from decimal import Decimal
+from datetime import datetime, timedelta
 
 
 @app.route('/')
@@ -95,32 +95,47 @@ def get_nutrition(ndbno):
             meal_choice = form1.meal.data
             quant_choice = int(form1.quantity.data)
 
-            food = Food(food_name=food_name, count=int(quant_choice), kcal=quant_choice*float(food_cals), protein=quant_choice*float(food_protein), fat=quant_choice*float(food_fat), carbs=quant_choice*float(food_carbs), unit=food_measure, meal=meal_choice, ndbno=ndbno, user_id=current_user.get_id())
+            food = Food(food_name=food_name, count=int(quant_choice), kcal=quant_choice * float(food_cals), protein=quant_choice * float(food_protein), fat=quant_choice * float(food_fat), carbs=quant_choice * float(food_carbs), unit=food_measure, meal=meal_choice, ndbno=ndbno, user_id=current_user.get_id())
             db.session.add(food)
             db.session.commit()
             return redirect(url_for('diary'))
 
 
 @app.route('/diary', methods=['GET', 'POST'])
-def diary():
+@app.route('/diary/<string:date_pick>', methods=['GET', 'POST'])
+def diary(date_pick=datetime.utcnow().strftime('%B %d, %Y')):
+
     form = RemoveFood()
-    foods = Food.query.filter_by(user_id=current_user.get_id())
+    form2 = DiaryDatePicker()
+
+    session.pop('_flashes', None)
+
+    form2.date.data = date_pick
+
+    foods = Food.query.filter_by(user_id=current_user.get_id(), date=date_pick)
 
     if request.method == 'POST':
-        remove_id = form.entry_id.data
-        user_id_for_row = Food.query.filter_by(id=remove_id)[0].user_id
-        if str(user_id_for_row) == current_user.get_id():
-            Food.query.filter_by(id=remove_id).delete()
-            db.session.commit()
-        else:
-            flash("Diary entry not found.")
-        return redirect(url_for('diary'))
+        if form.remove.data and form.validate():
+            remove_id = form.entry_id.data
+            user_id_for_row = Food.query.filter_by(id=remove_id)[0].user_id
+            if str(user_id_for_row) == current_user.get_id():
+                Food.query.filter_by(id=remove_id).delete()
+                db.session.commit()
+            else:
+                flash("Diary entry not found.")
+            return redirect(url_for('diary', date_pick=date_pick))
+        if form2.back.data and form2.validate():
+            date_pick = (datetime.strptime(form2.date.data, '%B %d, %Y') - timedelta(days=1)).strftime('%B %d, %Y')
+            return redirect(url_for('diary', date_pick=date_pick))
+        if form2.forward.data and form2.validate():
+            date_pick = (datetime.strptime(form2.date.data, '%B %d, %Y') + timedelta(days=1)).strftime('%B %d, %Y')
+            return redirect(url_for('diary', date_pick=date_pick))
 
     meals = ['Breakfast', 'Lunch', 'Dinner', 'Snacks']
-    total_cals = {'Breakfast':0, 'Lunch':0, 'Dinner':0, 'Snacks':0, 'total':0}
-    total_carbs = {'Breakfast':0, 'Lunch':0, 'Dinner':0, 'Snacks':0, 'total':0}
-    total_protein = {'Breakfast':0, 'Lunch':0, 'Dinner':0, 'Snacks':0, 'total':0}
-    total_fat = {'Breakfast':0, 'Lunch':0, 'Dinner':0, 'Snacks':0, 'total':0}
+    total_cals = {'Breakfast': 0, 'Lunch': 0, 'Dinner': 0, 'Snacks': 0, 'total': 0}
+    total_carbs = {'Breakfast': 0, 'Lunch': 0, 'Dinner': 0, 'Snacks': 0, 'total': 0}
+    total_protein = {'Breakfast': 0, 'Lunch': 0, 'Dinner': 0, 'Snacks': 0, 'total': 0}
+    total_fat = {'Breakfast': 0, 'Lunch': 0, 'Dinner': 0, 'Snacks': 0, 'total': 0}
 
     for food in foods:
         total_cals['total'] = total_cals['total'] + food.kcal
@@ -135,8 +150,7 @@ def diary():
                 total_protein[meal] = total_protein[meal] + food.protein
                 total_fat[meal] = total_fat[meal] + food.fat
 
-
-    return render_template('diary.html', foods=foods, form=form, total_fat=total_fat, total_cals=total_cals, total_carbs=total_carbs, total_protein=total_protein)
+    return render_template('diary.html', foods=foods, form=form, form2=form2, total_fat=total_fat, total_cals=total_cals, total_carbs=total_carbs, total_protein=total_protein)
 
 
 @app.route('/login', methods=['GET', 'POST'])
